Desacoplando etapas de cadastro e validando campos
