+++ src/StackExterna.tsx
@@ -12,6 +12,7 @@ import { useAuthContext } from "./util/context/providers/authProvider";
 import CriarProduto from "./components/App/Inicio/CriarProduto";
 import CriarMercado from "./components/App/Inicio/CriarMercado";
 import CriarSugestao from "./components/App/Inicio/CriarSugestao";
+import CadastroProvider from "./util/context/providers/cadastroProvider";
 
 export type StackExternaRoutesParams = {
 	login: undefined;
