+++ src/interfaces/models/Cadastro.ts
@@ -1,0 +1,5 @@
+import Usuario from "./Usuario";
+
+export default interface Cadastro extends Usuario {
+    senhaConfirma: string;
+}
